refactor: moved regions to supported-regions
