--- src/constants.ts
@@ -1,35 +1,0 @@
-// This list is extracted from https://docs.aws.amazon.com/general/latest/gr/rande.html
-// Tip: check git history to see when was the last time this list was updated
-export const awsRegionIds = [
-	'us-east-2',
-	'us-east-1',
-	'us-west-1',
-	'us-west-2',
-	'af-south-1',
-	'ap-east-1',
-	'ap-south-2',
-	'ap-southeast-3',
-	'ap-southeast-4',
-	'ap-south-1',
-	'ap-northeast-3',
-	'ap-northeast-2',
-	'ap-southeast-1',
-	'ap-southeast-2',
-	'ap-northeast-1',
-	'ca-central-1',
-	'ca-west-1',
-	'eu-central-1',
-	'eu-west-1',
-	'eu-west-2',
-	'eu-south-1',
-	'eu-west-3',
-	'eu-south-2',
-	'eu-north-1',
-	'eu-central-2',
-	'il-central-1',
-	'me-south-1',
-	'me-central-1',
-	'sa-east-1',
-	'us-gov-east-1',
-	'us-gov-west-1',
-] as const
